riscv: irqchip: plic: fix hunging in the plic_irq_resume() function
The address 'base+0x00' of plic is reserved, the loop
should start from 1 not 0.

In PLIC spec, Global interrupt sources are assigned small
unsigned integer identifiers, beginning at the value 1.
An interrupt ID of 0 is reserved to mean “no interrupt”.

The address 'base+0x00' might be illegal, so the loop should start from 1.

Signed-off-by: Jia Wang <wangjia@ultrarisc.com>
Signed-off-by: Yanteng Si <si.yanteng@linux.dev>