riscv: hwprobe: Export the Supm ISA extension
mainline inclusion
from mainline-v6.13
commit 3c2e0aff7b4f03fbc11b7d63c8db5b94a48978cf
category: feature
bugzilla: https://github.com/RVCK-Project/rvck-olk/issues/36

--------------------------------

Supm is a virtual ISA extension defined in the RISC-V Pointer Masking
specification, which indicates that pointer masking is available in
U-mode. It can be provided by either Smnpm or Ssnpm, depending on which
mode the kernel runs in. Userspace should not care about this
distinction, so export Supm instead of either underlying extension.

Hide the extension if the kernel was compiled without support for the
pointer masking prctl() interface.

Signed-off-by: Samuel Holland <samuel.holland@sifive.com>
Link: https://lore.kernel.org/r/20241016202814.4061541-9-samuel.holland@sifive.com
Signed-off-by: Palmer Dabbelt <palmer@rivosinc.com>
Signed-off-by: Weihao Li <ieiao@outlook.com>
Signed-off-by: Yanteng Si <si.yanteng@linux.dev>